requirements.txt
tornado==4.0.2
pdfminer.six
six
Tornado-JSON==0.41
python-Levenshtein==0.11.2
python-docx==0.7.4
pyzmq==14.4.0
PyMySQL==0.6.2
PyYAML==3.11

